Refactor HF dataset to always have numpy array instead of PIL image
